+++ analysis/src/main/kotlin/de/maibornwolff/dependacharta/pipeline/analysis/analyzers/delphi/DelphiAnalyzer.kt
@@ -1,0 +1,20 @@
+package de.maibornwolff.dependacharta.pipeline.analysis.analyzers.delphi
+
+import de.maibornwolff.dependacharta.pipeline.analysis.analyzers.BaseLanguageAnalyzer
+import de.maibornwolff.dependacharta.pipeline.analysis.model.FileInfo
+import de.maibornwolff.dependacharta.pipeline.analysis.model.Path
+import de.maibornwolff.dependacharta.pipeline.shared.SupportedLanguage
+import de.maibornwolff.treesitter.excavationsite.api.Declaration
+
+class DelphiAnalyzer(
+    fileInfo: FileInfo,
+) : BaseLanguageAnalyzer(fileInfo) {
+    override val language = SupportedLanguage.DELPHI
+
+    override fun buildPathWithName(
+        packagePath: List<String>,
+        declaration: Declaration,
+    ): Path {
+        return Path(packagePath + declaration.parentPath + declaration.name)
+    }
+}