+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add Delphi dependency analysis support (`.pas`, `.dpr` files)
+
 ## [0.23.0] - 2026-04-17
 
 ### Changed